+++ lib/widgets/transaction_filter_head/select_multi_account_sheet.dart
@@ -1,5 +1,6 @@
 import "package:flow/entity/account.dart";
 import "package:flow/l10n/extensions.dart";
+import "package:flow/utils/optional.dart";
 import "package:flow/utils/simple_query_sorter.dart";
 import "package:flow/widgets/general/flow_icon.dart";
 import "package:flow/widgets/general/frame.dart";
@@ -9,7 +10,7 @@ import "package:flutter/material.dart";
 import "package:go_router/go_router.dart";
 import "package:material_symbols_icons/symbols.dart";
 
-/// Pops with [List] of selected [Account]s
+/// Pops with an [Optional] of [List] of selected [Account]s
 class SelectMultiAccountSheet extends StatefulWidget {
   final List<Account> accounts;
   final List<String>? selectedUuids;
@@ -66,19 +67,31 @@ class _SelectMultiAccountSheetState extends State<SelectMultiAccountSheet> {
         alignment: .end,
         children: [
           TextButton.icon(
-            onPressed: () => context.pop(<Account>[]),
-            icon: const Icon(Symbols.block_rounded, fill: 0.0),
-            label: Text("transactions.query.clearSelection".t(context)),
-          ),
-          TextButton.icon(
             onPressed: pop,
             icon: const Icon(Symbols.check_rounded),
             label: Text("general.done".t(context)),
           ),
         ],
       ),
-      leading: showSearchBar
-          ? Frame(
+
+      // leading: showSearchBar
+      //     ? Frame(
+      //         child: TextField(
+      //           onChanged: (value) => setState(() => _query = value),
+      //           textInputAction: TextInputAction.done,
+      //           decoration: InputDecoration(
+      //             hintText: "general.search".t(context),
+      //             prefixIcon: const Icon(Symbols.search_rounded),
+      //           ),
+      //         ),
+      //       )
+      //     : null,
+      leading: Column(
+        mainAxisSize: .min,
+        spacing: 8.0,
+        children: [
+          if (showSearchBar)
+            Frame(
               child: TextField(
                 onChanged: (value) => setState(() => _query = value),
                 textInputAction: TextInputAction.done,
@@ -87,8 +100,22 @@ class _SelectMultiAccountSheetState extends State<SelectMultiAccountSheet> {
                   prefixIcon: const Icon(Symbols.search_rounded),
                 ),
               ),
-            )
-          : null,
+            ),
+          Frame(
+            child: Align(
+              alignment: AlignmentDirectional.topEnd,
+              child: TextButton(
+                onPressed: _toggleSelectAll,
+                child: Text(
+                  selectedUuids.length == widget.accounts.length
+                      ? "general.select.clear".t(context)
+                      : "general.select.all".t(context),
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
       child: SingleChildScrollView(
         child: Column(
           mainAxisSize: MainAxisSize.min,
@@ -129,11 +156,20 @@ class _SelectMultiAccountSheetState extends State<SelectMultiAccountSheet> {
     setState(() {});
   }
 
+  void _toggleSelectAll() {
+    if (selectedUuids.length == widget.accounts.length) {
+      selectedUuids.clear();
+    } else {
+      selectedUuids.addAll(widget.accounts.map((account) => account.uuid));
+    }
+    setState(() {});
+  }
+
   void pop() {
     final List<Account> selectedAccounts = widget.accounts
         .where((account) => selectedUuids.contains(account.uuid))
         .toList();
 
-    context.pop(selectedAccounts);
+    context.pop(Optional(selectedAccounts));
   }
 }